improve: 优化代码结构并通过 queryParameter 获取 imgKey
感觉这样可以利用好缓存的 imageUrl
