+++ src/zh/colamanga/build.gradle
@@ -5,7 +5,7 @@ ext {
     extName = 'Colamanga'
     pkgNameSuffix = 'zh.colamanga'
     extClass = '.Colamanga'
-    extVersionCode = 2
+    extVersionCode = 3
 }
 
 apply from: "$rootDir/common.gradle"
